backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.135.1",
    "motor>=3.7.1",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "python-magic-bin>=0.4.14",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = []
